#3930 added factory patterns to enable extending of env
