+++ environment/src/main/java/tygronenv/EntityListener.java
@@ -1,0 +1,19 @@
+package tygronenv;
+
+/**
+ * Listener for an entity
+ * 
+ * @author W.Pasman
+ *
+ */
+public interface EntityListener {
+
+	/**
+	 * Called by Entity when it is ready to give first percepts
+	 * 
+	 * @param entityName
+	 *            name for this entity.
+	 */
+	void entityReady(String entityName);
+
+}
